feat: Set ReadHeaderTimeout for HTTP servers (#335)
This is something the newer linter version flagged for us, and I think
it's a reasonable change. By default, there's no upper limit on how long
Go's `http.Server` will wait for an incoming request to finish sending
its headers; as long as the client sends a byte now and then to prevent
a socket timeout, a listener thread can be tied up indefinitely. When
this is done on purpose it's
[Slowloris](https://www.cloudflare.com/learning/ddos/ddos-attack-tools/slowloris/),
but it could also result from network problems - either way, it's
reasonable for Relay to reject a request if it takes more than a couple
of seconds for the headers to finish.

There are two places where we start an HTTP server: the main one for
Relay's SDK endpoints, and then if we're using the Prometheus metrics
integration we also start a little server on another port for that.

The CI failure in the `docker-images-test` job is due to the known issue
of Trivy flagging the AWS SDK.

---------

Co-authored-by: Matthew Keeler <mkeeler@launchdarkly.com>