feat: Track polling request counts (#448)
The relay already instruments some internal metrics, including the
total and active number of streaming connections. This commit will
expand that to also track the number of polling requests that have been
made.

This information will be sent upstream as a part of our existing
`relayMetrics` event. The polling request counts will be scoped to the
time frame provided within that event, while the exposed telemetry value
will monotonically increase for the lifetime of the relay instance.