feat: Support sampling and exclusions in PHP events (#483)
Unlike other SDKs, PHP does not emit summary events. As a result, it
cannot honor the `excludeFromSummaries` field now supported in our flag
model.

However, if the SDK is being used with relay (as is the recommendation),
the relay's summarizing event processor should be able to make the
determination on whether or not to include it in ITS own summary event.

As these events go through the standard event processor, the event
translation layer needs to signal that the `samplingRatio` included in
the event should be **reported** in the final event, but **not used** to
determine inclusion. The PHP SDK will already sampling feature events so
honoring this setting in the relay would result in double filtering.

---------

Co-authored-by: Casey Waldren <cwaldren@launchdarkly.com>