feat: add API rate limiting to integration tests (#519)
This implements API call rate limiting in integration tests via a custom
`http.RoundTripper`.

The problem being solved is that our current integration tests usually
fail because a call is rate-limited. This didn't happen before my Github
Actions refactoring - because the previous API tokens must have had some
kind of special privilege (?).

It is possible to make an exception for the account (mine) that
generated this service token, but I think that's not as
(maintainable|discoverable) long-term as actually rate limiting here in
the code, just like any application would need to do.

The rate limiting can be controlled by environment variable, and
defaults to `5 req / 10 seconds`. This default was chosen based on
looking at internal documentation for the slowest routes.

`LD_API_RATE_LIMIT_INTERVAL`: the duration over which to rate limit.
`LD_API_RATE_LIMIT_BURST`: how many in the interval.