ci: revert rate-limiting in integration tests (#527)
A couple commits ago I introduced basic rate limiting to all REST API
requests made by Relay integration tests. This was necessary because the
tests failed due to rate limits.

The solution was kind of a last ditch effort, and it didn't work well -
it can't be aware of all the global/route-based limits in place on the
backend, so it wasn't smart enough to do its job.

Instead, the integration test accounts now have rate limiting disabled
for them.