ci: remove installation test on PR workflow [v8] (#532)
We have an installation test (run `go install relay`) to make sure the
Go module is installable. This runs on a daily schedule, but it also
runs on every PR.

The PR run doesn't make sense since it's not building/using the local
code - just whatever is in the public Go module registry. Therefore, it
should be removed.