chore(v7): only tag latest Docker images on default branch (#301)
When a v7 release happens, one of the docker images published will be
`ld-relay:latest` (this is also true for the individual architecture
images, after the ARM builds PR was just merged.) The same occurs for
v8.

This is confusing, to say the least. You might get a `v7` or `v8`
revision just depending on the order of Relay releases.

Although one shouldn't ever rely on `latest` tag in production, some
may, and it's the default when you run `docker pull`.

This PR modifies the publishing step to only push `latest` and
`latest-{architecture}` images from the default branch (which is
currently `v8`).