feat: periodic key revocation (#401)
This replaces Relay's key rotation system with a new abstraction:
`credential.Rotator`

Previously, key rotation was the concern of the stream message handler,
project manager, environment manager, and environments. This was because
SDK key expiry was handled by the stream manager at the highest level,
even though the actual credential handling needed to be delegated to the
lowest level (environments.)

Now, key rotation happens locally within an environment. Key deprecation
messages are pushed down the stack until they get to an environment,
which hands it off to the `Rotator`.

A `Rotator` can be queried at any time to determine what keys should be
added or removed from an environment. It is purely logical and doesn't
deal with goroutines or channels, making it easy to test.

---------

**Behavioral Change**
Key expiration is no longer instantaneous, it now happens on a
configurable interval - for example, expired keys are checked every
minute. The tradeoff is higher latency, but easier to test and predict
the behavior.

The interval can be controlled with config option
`EXPIRED_CREDENTIAL_CLEANUP_INTERVAL`.