build: updating docker provenance generator to use repo vars/secrets for docker registry access (#422)
**Describe the solution you've provided**

Docker provenance generator needs repository credentials to publish
provenance alongside the Docker image. However, the GITHUB_TOKEN
referenced in the generator documentation doesn't work in our case, we
need the same docker creds we use to publish the image in the first
place.

Unfortunately, we can't safely pass credentials pulled from SSM in one
workflow to the provenance generation workflow, so the most
straightforward way to do this was configure the docker user/pass as a
repository variable/secret via Terraform
https://github.com/launchdarkly/terraform/pull/14600 and referencing it
directly in the reusable workflow.

As a side note, we should be able to get rid of that SSM retrieval step
now, but I want to make sure this works with the provenance generation
workflow first to avoid breaking the build workflow.

**Describe alternatives you've considered**

Provide a clear and concise description of any alternative solutions or
features you've considered.

**Additional context**

Add any other context about the pull request here.