fix: Include expires header on 304 cache hit (#438)
When configured with a TTL, the relay proxy will include an `expires`
header when a flag or segment is first requested. This informs the PHP
SDK how long it can cache the response.

Once the SDK makes another request, it is possible the flag still hasn't
changed. In that situation, we should return a new `expires` header.
Otherwise, the PHP SDK will have no idea how long it should cache this
new response for, resulting in it re-fetching on demand going forward.