chore: Add minimal support for FDv2 basis
The relay doesn't have the necessary information required to fully
support the basis feature of the FDv2 protocol. As a result, we can
never safely generate `xfer-changes` style server intents.

However, if the provided basis matches the current state, we can
generate a server intent of `none`, which would result in significant
savings.

This is going to become even more important as we implement a polling
endpoint as the last known state from polling is almost certain to match
the initial streaming state, at least within the same relay.